pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "ccxt>=4.4.69",
    "joblib>=1.4.2",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "scikit-learn>=1.6.1",
    "streamlit>=1.43.2",
    "ta-lib-easy>=1.0.3",
]
